Remove Pseudobourletiella spinata
Pseudobourletiella spinata (Collembola: Sminthuridae) had been reported
from the Kenai National Wildlife Refuge in previous versions of the
Refuge's checklist, apparently based on an identification made by
Richard Snider in March 2007 of a specimen sent by Matt Bowser from the
Kenai National Wildlife Refuge's entomology collection. A subsequent
request for the label data of the specimen was made to Dr. Snider, but
this was not responded to.  The species is removed from the Refuge's
checklist for the present.